+++ src/components/sidebar-elegible-items/elegible-list/ButtonElement.jsx
@@ -1,0 +1,22 @@
+import { WrapperComponents } from "../../../layouts/WrapperComponents";
+import MoldeElement from "./MoldeElement";
+
+const ButtonElement = () => {
+  return (
+    <WrapperComponents secctionName={"Buttons"}>
+      <MoldeElement
+        htmlType={"button"}
+        nameComponent={"Button"}
+        other={"button"}
+      />
+      <MoldeElement
+        htmlType={"a"}
+        nameComponent={"Link"}
+        other={"a"}
+        textToComponent="Link"
+      />
+    </WrapperComponents>
+  );
+};
+
+export default ButtonElement;