+++ src/components/sidebar-elegible-items/elegible-list/DivElement.jsx
@@ -1,4 +1,4 @@
-import { WrapperComponents } from '../../layouts/WrapperComponents'
+import { WrapperComponents } from '../../../layouts/WrapperComponents'
 import MoldeElement from './MoldeElement'
 
 const DivElement = () => {