+++ src/components/sidebarleftitems/SideBarLeftItems.jsx
@@ -1,44 +1,46 @@
-import { useState } from "react"
-import SideBarElementsItems from "../addcomponents/SideBarElementsItems"
-import SideBarElementsRendered from "../componentstree/SideBarElementsRendered"
-import LeftSideContainer from "../sidecontainer/LeftSideContainer"
-import { ListTreeIcon, PlusIcon } from "../svg"
-import ButtonRight from "../tooltips/ButtonRight"
+import { useState } from "react";
+import SideBarElementsRendered from "../componentstree/SideBarElementsRendered";
+import LeftSideContainer from "../sidecontainer/LeftSideContainer";
+import { ListTreeIcon, PlusIcon } from "../svg";
+import ButtonRight from "../tooltips/ButtonRight";
+import SideBarElementsItems from "../sidebar-elegible-items/addcomponents/SideBarElementsItems";
 
 const SideBarLeftItems = () => {
-  const [openComp, setOpenComp] = useState({ type: "" })
+  const [openComp, setOpenComp] = useState({ type: "" });
 
   const handleShowComponents = (type) => {
-    if (openComp.type == type)
-      setOpenComp({ type: "" })
-    else
-      setOpenComp({ type })
-  }
+    if (openComp.type == type) setOpenComp({ type: "" });
+    else setOpenComp({ type });
+  };
 
   return (
     <>
-      {openComp.type != "" &&
+      {openComp.type != "" && (
         <LeftSideContainer handleShowComponents={handleShowComponents}>
-          {openComp.type == "addcomp" &&
-            <SideBarElementsItems />
-          }
-          {openComp.type == "tree" &&
-            <SideBarElementsRendered />
-          }
+          {openComp.type == "addcomp" && <SideBarElementsItems />}
+          {openComp.type == "tree" && <SideBarElementsRendered />}
         </LeftSideContainer>
-      }
+      )}
       <div>
-        <ButtonRight text={"Add Component"} handleShowComponents={handleShowComponents} type={"addcomp"}>
+        <ButtonRight
+          text={"Add Component"}
+          handleShowComponents={handleShowComponents}
+          type={"addcomp"}
+        >
           <PlusIcon />
         </ButtonRight>
       </div>
       <div>
-        <ButtonRight text={"Components Tree"} handleShowComponents={handleShowComponents} type={"tree"}>
+        <ButtonRight
+          text={"Components Tree"}
+          handleShowComponents={handleShowComponents}
+          type={"tree"}
+        >
           <ListTreeIcon />
         </ButtonRight>
       </div>
     </>
-  )
-}
+  );
+};
 
-export default SideBarLeftItems
+export default SideBarLeftItems;